+++ lib/commanded/eventstore/event_store.ex
@@ -28,9 +28,9 @@ defmodule Commanded.EventStore do
 
   @callback read_snapshot(String.t) :: {:ok, SnapshotData.t} | {:error, :snapshot_not_found}
   
-  @callback read_stream_forward(String.t) :: {:ok, list(RecordedEvent.t)} | {:error, reason :: term}
-  @callback read_stream_forward(String.t, non_neg_integer) :: {:ok, list(RecordedEvent.t)} | {:error, reason :: term}
   @callback read_stream_forward(String.t, non_neg_integer, non_neg_integer) :: {:ok, list(RecordedEvent.t)} | {:error, reason :: term}
+
+  @callback stream_forward(String.t, non_neg_integer, non_neg_integer) :: Enumerable.t | {:error, reason :: term}
   
   @callback subscribe_to_all_streams(String.t, pid, start_from) :: {:ok, subscription :: any}
     | {:error, :subscription_already_exists}
@@ -40,8 +40,6 @@ defmodule Commanded.EventStore do
 
   @callback record_snapshot(SnapshotData.t) :: :ok | {:error, reason :: term}
 
-  @callback read_all_streams_forward() :: {:ok, list(RecordedEvent.t)} | {:error, reason :: term}
-  @callback read_all_streams_forward(non_neg_integer) :: {:ok, list(RecordedEvent.t)} | {:error, reason :: term}
   @callback read_all_streams_forward(non_neg_integer, non_neg_integer) :: {:ok, list(RecordedEvent.t)} | {:error, reason :: term}
 
   @callback delete_snapshot(String.t) :: :ok | {:error, reason :: term}